add sparse sgd in Go (#1679)
* add embedding table struct in Go

* add sparse sgd kernel

* add new embedding table func

* fix ci

* fix ci

* add more test for SGD optimizer

+++ elasticdl/pkg/kernel/capi/kernel_api.cc
@@ -3,7 +3,7 @@
 #include <cmath>
 #include <eigen3/Eigen/Dense>
 
-void SGD(float* grad, float* param, double lr, long long size) {
+void SGD(float* grad, float* param, float lr, long long size) {
   Eigen::Map<Eigen::Array<float, 1, Eigen::Dynamic>> eg{
       grad, static_cast<Eigen::Index>(size)};
